Movement of file manipulation to own file within stringHandling
