+++ src/main/kotlin/stringHandling/display.kt
@@ -3,8 +3,7 @@ package stringHandling
 import api.Password
 import encrytion.aesDecrypt
 import encrytion.secKey
-import java.io.BufferedReader
-import java.io.File
+
 
 fun largest(list: List<String>): Int {
     //this works by adding to a list and then sorting, reversing and taking the first value
@@ -83,13 +82,4 @@ fun displayAsTable(data: List<Password>): String {
     }
     return (stringOut)
 
-}
-fun readFromFile(file: String): String {
-    //this is for reading the key
-    val bufferedReader: BufferedReader = File("src/main/resources/$file").bufferedReader()
-    val inputString = bufferedReader.use { it.readText() }
-    return (inputString)
-}
-fun writeToFile(file: String, text: String) {
-    File("src/main/resources/$file").writeText(text)
 }
